refactor: CodeSpectator 코드 영역 분리로 리렌더 최소화 (#261) - CodeSpectatorEditor로 코드/진행률 구독 분리 - codeupdate 시 코드 영역만 리렌더되도록 개선
